+++ attributes/default.rb
@@ -3,11 +3,9 @@ default['wade']['web']['postgres_user'] = 'postgres'
 default['wade']['web']['postgres_addr'] = 'localhost'
 default['wade']['web']['postgres_db'] = 'WADE'
 default['wade']['web']['hostname'] = node['hostname']
-default['wade']['web']['version'] = '0.0.1-1'
+default['wade']['web']['version'] = '0.0.1-7'
 default['wade']['web']['ssl']['active'] = true
-default['wade']['web']['ssl']['cert']['location'] = '/tmp/kitchen/data/example.com.cert'
-default['wade']['web']['ssl']['key']['location'] = '/tmp/kitchen/data/example.com.key'
+default['wade']['web']['ssl']['cert']['location'] = 'file:///tmp/kitchen/data/example.com.cert'
+default['wade']['web']['ssl']['key']['location'] = 'file:///tmp/kitchen/data/example.com.key'
 default['wade']['yum']['repo']['baseurl'] = 'https://cida.usgs.gov/artifactory/rpm'
-
-
 default['wade']['databag']['databag_name'] = 'wade-_default'